# documentation/open-api.spec.yml
openapi: 3.1.0
info:
  title: Ecommerce
  description: |-
    This is the API documentation for our general purpose ecommerce REST API

    Some useful links:
    - [RESTful API repository](https://github.com/ilopez-ravn/ecommerce)


  contact:
    email: israellopez@ravn.co
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8080/api/v1
  - url: https://ecommerce-api.pacasoft.app/api/v1
...
security:
  - bearerAuth: []  
paths:
...
  /warehouses:
    get:
      tags:
        - warehouses
      summary: Get all warehouses
      description: Retrieve a list of all warehouses
      operationId: getWarehouses
      responses:
        '200':
          description: List of warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Warehouse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - warehouses
      summary: Create a new warehouse
      description: Create a new warehouse
      operationId: createWarehouse
      requestBody:
        description: Warehouse object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostWarehouse'
        required: true
      responses:
        '201':
          description: Warehouse created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
                
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /warehouses/{warehouseId}:
    put:
      tags:
        - warehouses
      summary: Update an existing warehouse
      description: Update an existing warehouse by its unique identifier
      operationId: updateWarehouse
      parameters:
        - name: warehouseId
          in: path
          description: ID of warehouse to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Warehouse object that needs to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostWarehouse'
        required: true
      responses:
        '200':
          description: Warehouse updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
                
        '404':
          description: Warehouse not found
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      

    delete:
      tags:
        - warehouses
      summary: Delete a warehouse by ID
      description: Delete a single warehouse by its unique identifier
      operationId: deleteWarehouseById
      parameters:
        - name: warehouseId
          in: path
          description: ID of warehouse to delete
          required: true
          schema:
            type: integer
            format: int64
      
      responses:
        '204':
          description: Warehouse deleted successfully
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
                
        '404':
          description: Warehouse not found
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /warehouses/{warehouseId}/stock:
    post:
      tags:
        - warehouses
      summary: Update stock levels for multiple products
      description: Update stock levels for multiple products
      operationId: updateProductStocks
      requestBody:
        description: List of product stock updates
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductStockUpdateList'
        required: true
      responses:
        '200':
          description: Product stocks updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Product stocks updated successfully
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
                
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Warehouse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Main Warehouse"
        location:
          type: string
          example: "123 Warehouse St, City, Country"
    PostWarehouse:
      type: object
      required: 
        - name
        - location
      properties:
        name:
          type: string
          example: "Main Warehouse"
        location:
          type: string
          example: "123 Warehouse St, City, Country"
    ProductStockUpdateList:
      type: object
      required: 
        - product_id
        - type
        - quantity
      properties:
        product_id:
          type: integer
          format: int64
        type:
          type: string
          enum:
            - ADD
            - SUBTRACT
        quantity:
          type: integer
...
    Error:
      type: object
      properties:
        code:
          type: number
          example: 500
        message:
          type: string
          example: "Internal server error"
      required:
        - code
        - message
...
    InvalidQueryResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 400
        message:
          type: string
          example: "Invalid query parameter"
      required:
        - code
        - message
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes